13.1 & 13.2
part 2 was a stumper. started brute-force knowing it wouldn't be
efficient enough but it was kind of fun spin up doing billions of
arithmetic ops. finally a had a visual of aligning sliding locks and
realized i could start building necessary but not sufficient intervals
and grow the answer piece by piece.